refactor: Restructure Tetris component files
